chore: moved structure to Layout component
