+++ src/pages/ui-components/Layout/Sidebar/MenuItem/index.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { cx } from '@emotion/css';
 import { useContextSelector } from 'use-context-selector';
 
-import styled, { IStyledProp } from '../../styles/styled';
+import styled, { IStyledProp } from '../../../styles/styled';
 import SidebarContext from '../context';
 import SubMenuItemContext from '../SubMenuItem/context';
